Added modal dialogs
